requirements.txt
tweepy
openai
python-dotenv
requests